fix(xcm): get rid of xcm warns, fix quartz CurrencyIdConvert
