+++ runtime/opal/src/tests/mod.rs
@@ -14,5 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with Unique Network. If not, see <http://www.gnu.org/licenses/>.
 
-mod xcm;
 mod logcapture;
+mod xcm;
